Refactor configuration files and data handling
- Updated `ccf_config.yaml` to add GPIO configurations and remove redundant entries.
- Modified `config.yml` to adjust sensor calibration points and reorganize channel IDs for consistency.
- Created a new `test_config.yaml` for testing purposes with default values for sensors and actuators.
- Implemented `data_interface.py` to handle data processing and saving to CSV files.
- Enhanced `main.py` to integrate new data saving functionality and toggle saving state.
- Updated `mqtt_interface.py` to streamline data processing and remove unused functions.
- Added JavaScript functionality in `main.js` for downloading data files.
- Updated `index.html` to include a button for downloading data.
