+++ backend/app/mqtt_interface.py
@@ -4,16 +4,20 @@ import json
 from datetime import datetime
 import asyncio
 import time
-from config_parser import get_config, process_data, convert_command, save_data, SAVE_DATA_FLAG
+import config_parser
+import data_interface
+
 MQTT_BROKER = "host.docker.internal" # use broker.hivemq.com for testing on PCs
 MQTT_PORT = 1883 # TCP Port
 DATA_TOPIC = "novaground/telemetry"
 COMMAND_TOPIC = "novaground/command"
 
 raw_data = {}
-data_store = {"sensors": [], "actuators": []}
 processed_data = {"sensors": [], "actuators": [], "gpios": []}
+data_store = []
 processed_gpios = []
+
+
 # MQTT client setup
 mqtt_client = mqtt.Client()
 
@@ -39,20 +43,7 @@ def on_message(client, userdata, msg):
             # print(f"Decoded data: {data}"
             # Make sure payload is a dictionary before using it in the process
             if isinstance(raw_data, dict):
-                if "sensors" in raw_data:
-                    processed_data["sensors"] = asyncio.run(process_data(raw_data, "sensors"))
-                    
-                if "gpios" in raw_data:
-                    processed_data["gpios"] = asyncio.run(process_data(raw_data, "gpios"))
-
-                if "actuators" in raw_data:
-                    processed_data["actuators"] = asyncio.run(process_data(raw_data, "actuators"))
-
-                if "thermocouples" in raw_data:
-                    processed_data["thermocouples"] = asyncio.run(process_data(raw_data, "thermocouples"))
-
-                if SAVE_DATA_FLAG:
-                    save_data(processed_data)  # Save the processed data to a file
+                asyncio.run(data_interface.process_data(raw_data))
             else:
                 print("Received payload is not a valid dictionary")
         else:
@@ -79,181 +70,9 @@ async def process_mqtt_message(payload):
 
 async def publish_command(command):
     try:
-        mqtt_client.publish(mqtt.COMMAND_TOPIC, json.dumps(command))
+        mqtt_client.publish(COMMAND_TOPIC, json.dumps(command))
         return {"status": "Command sent"}
     except Exception as e:
         print(f"Error publishing command: {e} - Payload: {json.dumps(command)}")
         raise Exception(f"Error publishing command: {e} - Payload: {json.dumps(command)}")
     
-
-async def set_all_to_closed():
-    """Set all servos and solenoids to their closed state."""
-    config = get_config()
-    commands = []
-    """
-    # Send commands to close all solenoids
-    for relay in config["relays"].values():
-        if relay.get("type") is None:
-            continue
-        command = {
-            "type": "relay",
-            "name": relay["name"],
-            "state": "closed"
-        }
-        relay_commands = await convert_command(command)
-        commands.extend(relay_commands)
-    # Send commands to close all servos
-    for servo in config["servos"].values():
-        command = {
-            "type": "servo",
-            "name": servo["name"],
-            "state": "closed"
-        }
-        servo_commands = await convert_command(command)
-        commands.extend(servo_commands)
-    # Publish all commands to the MQTT broker
-    for command in commands:
-        await publish_command(command)
-        time.sleep(0.1)  # Add a small delay between commands to avoid flooding the broker
-    """
-    for i in range(16):
-        command = {
-            "type": "relay",
-            "id": i,
-            "state": 1
-        }
-        mqtt_client.publish(mqtt.COMMAND_TOPIC, json.dumps(command))
-        time.sleep(0.1)  # Add a small delay between commands to avoid flooding the broker
-    return {"status": "Commands sent"}
-
-async def set_to_defaults():
-    """Set all actuators to their default state."""
-    config = get_config()
-    commands = []
-    wait = {"wait"}
-    for servo in config["servos"].values():
-        # Send commands to set all servos to their default state
-        name = servo["name"]
-        default_state = servo["default_state"]
-
-        if default_state is None:
-            raise ValueError(f"Default state not defined for servo '{name}'.")
-                
-        if default_state == "open":
-            angle = servo["open_pos"]
-        elif default_state == "closed":
-            angle = servo["close_pos"]
-            
-        command = {
-        "type": "servo",
-        "id": servo["channelID"],
-        "angle": angle
-        }
-        if "relayID" in servo:
-            power_on = {
-                "type": "relay",
-                "id": servo["relayID"],
-                "state": 0
-            }
-            power_off = {
-                "type": "relay",
-                "id": servo["relayID"],
-                "state": 1
-            }
-        
-        if power_on: 
-            commands.append(power_on)
-            commands.append(wait)
-        
-        commands.append(command)
-
-        if power_off: 
-            commands.append(wait)
-            commands.append(power_off)
-        
-    for gpio in config["gpios"].values():
-        # Send commands to set all GPIOs to their default state
-        name = gpio["name"]
-        default_state = gpio["default_state"]
-
-        if default_state is None:
-            raise ValueError(f"Default state not defined for GPIO '{name}'.")
-
-        if default_state == "armed":
-            state = 1
-        elif default_state == "disarmed":
-            state = 0
-        
-        if "relayID" in gpio:
-            power_on = {
-                "type": "relay",
-                "id": gpio["relayID"],
-                "state": 0
-            }
-            power_off = {
-                "type": "relay",
-                "id": gpio["relayID"],
-                "state": 1
-            }
-        command = {
-            "type": "gpio",
-            "id": gpio["channelID"],
-            "state": state
-        }
-        if power_on: 
-            commands.append(power_on)
-            commands.append(wait)
-        
-        commands.append(command)
-
-        if power_off: 
-            commands.append(wait)
-            commands.append(power_off)
-
-    for relay in config["relays"].values():
-        # Send commands to set all solenoids to their default state
-        if relay["actuator_type"] == "solenoid":
-            name = relay["name"]
-            relay_type = relay["relay_type"]
-            solenoid_type = relay["solenoid_type"]
-            default_state = relay["default_state"]
-
-            if relay_type not in ["NO", "NC"]:
-                raise ValueError(f"Invalid relay type '{relay_type}' for relay '{name}'.")
-            if solenoid_type not in ["NO", "NC"]:
-                raise ValueError(f"Invalid solenoid type '{solenoid_type}' for relay '{name}'.")
-            if default_state not in ["open", "closed"]:
-                raise ValueError(f"Invalid default state '{default_state}' for relay '{name}'.")
-            
-            power_state = "on" if (default_state == "open" and solenoid_type == "NC") or (default_state == "closed" and solenoid_type == "NO") else "off"
-            relay_state = 0 if (power_state == "on" and  relay_type == "NO" ) or (power_state == "off" and relay_type == "NC") else 1
-            command = {
-                "type": "relay",
-                "id": relay["channelID"],
-                "state": relay_state
-            }
-            commands.append(command)
-        if relay["actuator_type"] == "poweredDevice":
-            name = relay["name"]
-            relay_type = relay["relay_type"]
-            default_state = relay["default_state"]
-
-            if relay_type not in ["NO", "NC"]:
-                raise ValueError(f"Invalid relay type '{relay_type}' for relay '{name}'.")
-            if default_state not in ["on", "off"]:
-                raise ValueError(f"Invalid default state '{default_state}' for relay '{name}'.")
-
-            relay_state = 0 if (default_state == "on" and  relay_type == "NO" ) or (default_state == "off" and relay_type == "NC") else 1
-            command = {
-                "type": "relay",
-                "id": relay["channelID"],
-                "state": relay_state
-            }
-            commands.append(command)
-    
-    for command in commands:
-        if command == wait:
-            time.sleep(2)
-        else:
-            mqtt_client.publish(mqtt.COMMAND_TOPIC, json.dumps(command))
-            time.sleep(0.1)  # Add a small delay between commands to avoid flooding the broker
