+++ backend/app/templates/index.html
@@ -11,6 +11,7 @@
   <p>This is a simple interface for interacting with the backend services.</p>
   <button id="update-button">Update Config</button>
   <button id="save-button">Start saving data</button>
+  <button id="download-button">Download data</button>
   <button id="lock-button">Lock Actuators</button>
   <button id="calibration-button">Get Uncalibrated values</button>
   <p id="update-status"></p>
